Update apps.json to new AltSource API
At the moment, I've left the original API for keeping app details in there to support AltStore Beta versions below 1.6, but the properties can be removed at a later date. 

For now, I would also recommend creating a new "version" anytime there is a new release as well to maintain old versions of the app in perpetuity. I haven't tested with the latest AltStore Beta, but there was an oddity where AltStore v1.6b3 would only show the first version in the list regardless of which version was higher, so keep this in mind.